Default to template field ID for parsed issue body key
By default, when an issue template is parsed, the action will now attempt to match the header in the issue body markdown with the label field of the issue template. If found, it will check if an ID is present for that field. If present, the ID will be used for the property key. Otherwise, the action reverts to the original behavior (slugifying the field label).

E.g. If the field looks like this in the template:

```yaml
  - type: input
    id: name
    attributes:
      label: The Name of the Thing
      description: The name of the thing you want to create.
      placeholder: this-is-the-thing
    validations:
      required: true
```

Then the parsed issue body will loo like this:

```json
{
  "name": "Whatever the user-provided value is"
}
```

Otherwise, if the field looks like this in the template

```yaml
  - type: input
    attributes:
      label: The Name of the Thing
      description: The name of the thing you want to create.
      placeholder: this-is-the-thing
    validations:
      required: true
```

Then the parsed issue body JSON will look like:

```json
{
  "the_name_of_the_thing": "Whatever the user-provided value is"
}
```